requirements.txt
pandas>=1.0.3
pyyaml>=5.3.1
argparse>=1.1
rootpath>=0.1.1
anndata>=0.7.1
requests>=2.23.0